+++ src/migrations/2019_11_06_124901_create_taxes_table.php
@@ -1,44 +1,43 @@
 <?php
 
-use Illuminate\Support\Facades\Schema;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
-class CreateTaxesTable extends Migration
-{
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::create('taxes', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('company_id');
-            $table->string('name',191);
-            $table->unsignedInteger('type_id')->nullable();
-            $table->unsignedInteger('created_by_id')->nullable();
-            $table->unsignedInteger('updated_by_id')->nullable();
-            $table->unsignedInteger('deleted_by_id')->nullable();
-            $table->timestamps();
-            $table->softDeletes();
-            $table->foreign('company_id')->references('id')->on('companies')->onDelete('CASCADE')->onUpdate('cascade');
-            $table->foreign('type_id')->references('id')->on('configs')->onDelete('SET NULL')->onUpdate('cascade');
-            $table->foreign('created_by_id')->references('id')->on('users')->onDelete('SET NULL')->onUpdate('cascade');
-            $table->foreign('updated_by_id')->references('id')->on('users')->onDelete('SET NULL')->onUpdate('cascade');
-            $table->foreign('deleted_by_id')->references('id')->on('users')->onDelete('SET NULL')->onUpdate('cascade');
-            $table->unique(["company_id","name"]);
-        });
-    }
+class CreateTaxesTable extends Migration {
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up() {
+		if (!Schema::hasTable('taxes')) {
+			Schema::create('taxes', function (Blueprint $table) {
+				$table->increments('id');
+				$table->unsignedInteger('company_id');
+				$table->string('name', 191);
+				$table->unsignedInteger('type_id')->nullable();
+				$table->unsignedInteger('created_by_id')->nullable();
+				$table->unsignedInteger('updated_by_id')->nullable();
+				$table->unsignedInteger('deleted_by_id')->nullable();
+				$table->timestamps();
+				$table->softDeletes();
+				$table->foreign('company_id')->references('id')->on('companies')->onDelete('CASCADE')->onUpdate('cascade');
+				$table->foreign('type_id')->references('id')->on('configs')->onDelete('SET NULL')->onUpdate('cascade');
+				$table->foreign('created_by_id')->references('id')->on('users')->onDelete('SET NULL')->onUpdate('cascade');
+				$table->foreign('updated_by_id')->references('id')->on('users')->onDelete('SET NULL')->onUpdate('cascade');
+				$table->foreign('deleted_by_id')->references('id')->on('users')->onDelete('SET NULL')->onUpdate('cascade');
+				$table->unique(["company_id", "name"]);
+			});
+		}
+	}
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        Schema::dropIfExists('taxes');
-    }
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down() {
+		Schema::dropIfExists('taxes');
+	}
 }
